FEAT Enhances component titles for SEO
Improves SEO by conditionally rendering H1 tags instead of spans for main titles in several components.

A new `h1Choice` prop is introduced to components, allowing pages to specify if a H1 tag should be used instead of a span for the main title. This enables the use of meaningful H1 titles on specific pages to boost SEO.